+++ jaot_stock/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 # License LGPL-3
+from . import jaot_scenario
 from . import stock_picking